config-file Move getting options to separate folder
