+++ src/lint-files/__tests__/index.spec.ts
@@ -2,7 +2,7 @@ jest.mock('../lint-file');
 
 import lintFile from '../lint-file';
 import lintFiles from '../index';
-import { Formats } from '../../get-options';
+import { Formats } from '../../options';
 
 const mockedLintFile = (lintFile as unknown) as jest.Mock<string | null>;
